requirements.txt
opencv-python==3.4.1.15
Pillow==5.1.0
scipy==1.1.0
sklearn==0.0
tensorflow-gpu==1.8.0
